+++ test/event_tests/Event_test.cpp
@@ -15,10 +15,9 @@ using namespace owl::event;
 TEST(EventDispatcher, base) {
 	AppTickEvent event;
 	EventDispatcher dispatcher(event);
-	EXPECT_FALSE(
-			dispatcher.dispatch<AppUpdateEvent>([](Event &) { return true; }));
-	EXPECT_TRUE(dispatcher.dispatch<AppTickEvent>([](Event &) { return false; }));
+	EXPECT_FALSE(dispatcher.dispatch<AppUpdateEvent>([](Event&) { return true; }));
+	EXPECT_TRUE(dispatcher.dispatch<AppTickEvent>([](Event&) { return false; }));
 	EXPECT_FALSE(event.handled);
-	EXPECT_TRUE(dispatcher.dispatch<AppTickEvent>([](Event &) { return true; }));
+	EXPECT_TRUE(dispatcher.dispatch<AppTickEvent>([](Event&) { return true; }));
 	EXPECT_TRUE(event.handled);
 }