+++ test/renderer_tests/TextureAsync_test.cpp
@@ -7,9 +7,9 @@
  */
 #include "testHelper.h"
 
+#include "core/task/Scheduler.h"
 #include "renderer/gpu/RenderCommand.h"
 #include "renderer/gpu/Texture.h"
-#include "core/task/Scheduler.h"
 
 using namespace owl::renderer;
 using namespace owl::renderer::gpu;